+++ .idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="..\:/Projects/RMALV/AtosZadatak15/app/src/main/res/layout/activity_main.xml" value="0.358695652173913" />
         <entry key="..\:/Projects/administration-atos/app/src/main/res/layout/activity_main.xml" value="0.358695652173913" />
+        <entry key="..\:/Projects/administration-atos/app/src/main/res/layout/fragment_add_new_task.xml" value="0.358695652173913" />
         <entry key="..\:/Projects/administration-atos/app/src/main/res/layout/fragment_home.xml" value="0.358695652173913" />
         <entry key="..\:/Projects/administration-atos/app/src/main/res/layout/task_item.xml" value="0.358695652173913" />
       </map>
