+++ app/src/main/java/com/mesinger/atoszadatak15/ui/AddNewTaskFragment.java
@@ -1,27 +1,27 @@
 package com.mesinger.atoszadatak15.ui;
 
 import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import androidx.fragment.app.Fragment;
+import com.mesinger.atoszadatak15.R;
 
-import com.mesinger.atoszadatak15.databinding.FragmentAddNewTaskBinding;
+public class AddNewTaskFragment extends Fragment {
 
-class AddNewTaskFragment extends Fragment {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
 
-
-
-    private FragmentAddNewTaskBinding binding;
-
+    }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        binding = FragmentAddNewTaskBinding.inflate(inflater, container, false);
-        View view = binding.getRoot();
-        return view;
-
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_add_new_task, container, false);
     }
 }
