+++ app/src/main/res/layout/fragment_add_new_task.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.HomeFragment">
+    tools:context=".ui.AddNewTaskFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
